fix: beta tester feedback — remove fake stats, 16PF refs, add mobile responsive, date formatting
Landing page:
- Remove fake numbers (2000+ patients, 500+ psychologists, 300K sessions)
- Replace with qualitative trust badges and real value propositions
- Remove all 16PF test references from i18n (es/en) and components
- Replace "P" avatar in video mockup with friendly psychologist SVG
- Add security tags to VideoChat (AES-256-GCM, RGPD, Servidores EU)
- Add "Saber más" button to ForProfessionals section
- Add inline demo request form to ForClinics section
- Generalize TestsPreview: generic radar labels, bipolar bars, test type cards
- Update SmartMatching stats to generic algorithm references

SoyProfesional:
- Replace fake stats with real ones (24/7 soporte, 0% comisión)
- Replace "Asistente IA" with "Tareas personalizadas"
- Add "Consentimientos informados" and "Trae tus pacientes" cards
- Move COP badge section after main CTA
- Remove 16PF from SEO description

PsychDashboard mobile:
- Add hamburger menu + sidebar toggle with overlay
- Add mobile bottom navigation bar (Inicio, Pacientes, Agenda, Chat)
- Make content padding responsive, add close button to sidebar

Other fixes:
- Date formatting: month 'short' → 'long' across 25 components
- Add 12 new specializations and extended experience ranges
- Hide psychologist email from patient view
- Add consent status badges to patient profile header
- Remove obsolete MUSTS_ANTES_DE_DEPLOY.md and ROADMAP.md

+++ frontend/src/components/ChatWidget.tsx
@@ -782,7 +782,7 @@ export default function ChatWidget({ mode, otherId }: Props) {
                 <p className="text-sm text-slate-800">
                   {new Date(patientContext.lastAppointment).toLocaleDateString('es-ES', {
                     day: '2-digit',
-                    month: 'short',
+                    month: 'long',
                     year: 'numeric',
                   })}
                 </p>
@@ -794,7 +794,7 @@ export default function ChatWidget({ mode, otherId }: Props) {
                 <p className="text-sm text-slate-800">
                   {new Date(patientContext.assignedAt).toLocaleDateString('es-ES', {
                     day: '2-digit',
-                    month: 'short',
+                    month: 'long',
                     year: 'numeric',
                   })}
                 </p>
